modification table produits, controlleurs

+++ database/migrations/2024_01_10_204250_create_achats_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->integer('produit_id')->nullable();
             $table->integer('fournisseur_id')->nullable();
             $table->integer('user_id')->nullable();
-            $table->integer('quantite')->nullable();
-            $table->integer('prix_unitaire')->nullable();
-            $table->integer('prix_total')->nullable();
+            $table->double('quantite')->nullable(); 
+            $table->double('prix_unitaire')->nullable();
+            $table->double('prix_total')->nullable();
             $table->date('date_achat')->nullable();
             $table->string('statut')->nullable();
             $table->string('archive')->default(false);
